refactors PokemonsDetails component
