+++ src/components/Box/index.tsx
@@ -13,7 +13,7 @@ export interface BoxProps extends HTMLMotionProps<'div'> {
   $flexWrap?: string | Record<string, any>;
   width?: string | Record<string, any>;
   height?: string | Record<string, any>;
-  minHeight?: string | Record<string, any>;
+  $minHeight?: string | Record<string, any>;
   direction?: string | Record<string, any>;
   align?: string | Record<string, any>;
   justify?: string | Record<string, any>;
