+++ src/components/Type/TypeIcon/index.js
@@ -1,26 +1,25 @@
-import React from 'react'
+import React from 'react';
 // components
-import Box from '../../Box'
-import TypeBadge from '../../TypeBadge'
+import Box from '../../Box';
+import TypeBadge from '../../TypeBadge';
 // styles
-import { JpnName } from '../../BaseStyles'
+import { JpnName } from '../../BaseStyles';
 
 export default function TypeIcon({ typeName, otherNames, ...rest }) {
-  const japanName = otherNames.find(name => name.language.name === 'ja-Hrkt')
-    .name
+  const japanName = otherNames.find(name => name.language.name === 'ja-Hrkt').name;
 
   return (
-    <Box relative minHeight={{ xxs: '250px', lg: '350px' }} {...rest}>
+    <Box relative $minHeight={{ xxs: '250px', lg: '350px' }} {...rest}>
       <TypeBadge
-        type={typeName}
+        typename={typeName}
         key={`type-icon-${typeName}`}
         $iconOnly
-        fill="true"
+        fill={true}
         $float
         $iconWidth="auto"
         $iconHeight="150px"
       />
       {japanName && <JpnName>{japanName}</JpnName>}
     </Box>
-  )
+  );
 }
